+++ tests/cards/leaders/Bjorn.spec.ts
@@ -3,7 +3,7 @@ import {Game} from '../../../src/server/Game';
 import {TestPlayer} from '../../TestPlayer';
 import {forceGenerationEnd} from '../../TestingUtils';
 import {getTestPlayer, newTestGame} from '../../TestGame';
-import {Bjorn} from '../../../src/server/cards/leaders/Bjorn';
+import {Bjorn} from '../../../src/server/cards/ceos/Bjorn';
 import {LawSuit} from '../../../src/server/cards/promo/LawSuit';
 
 
@@ -17,7 +17,7 @@ describe('Bjorn', function() {
 
   beforeEach(() => {
     card = new Bjorn();
-    game = newTestGame(4, {leadersExtension: true});
+    game = newTestGame(4, {ceoExtension: true});
     player = getTestPlayer(game, 0);
     player2 = getTestPlayer(game, 1);
     player3 = getTestPlayer(game, 2);
