+++ tests/cards/leaders/Ulrich.spec.ts
@@ -3,7 +3,7 @@ import {Game} from '../../../src/server/Game';
 import {forceGenerationEnd, maxOutOceans} from '../../TestingUtils';
 import {getTestPlayer, newTestGame} from '../../TestGame';
 import {TestPlayer} from '../../TestPlayer';
-import {Ulrich} from '../../../src/server/cards/leaders/Ulrich';
+import {Ulrich} from '../../../src/server/cards/ceos/Ulrich';
 
 describe('Ulrich', function() {
   let card: Ulrich;
@@ -13,7 +13,7 @@ describe('Ulrich', function() {
 
   beforeEach(() => {
     card = new Ulrich();
-    game = newTestGame(2, {leadersExtension: true});
+    game = newTestGame(2, {ceoExtension: true});
     player = getTestPlayer(game, 0);
     player2 = getTestPlayer(game, 1);
   });
